Move provider passport projection write port wiring config
